net/neighbour: queue work on power efficient wq
Workqueue used in neighbour layer have no real dependency of scheduling
these on
the cpu which scheduled them.

On a idle system, it is observed that an idle cpu wakes up many times
just to
service this work. It would be better if we can schedule it on a cpu
which the
scheduler believes to be the most appropriate one.

This patch replaces normal workqueues with power efficient versions.
This
doesn't change existing behavior of code unless
CONFIG_WQ_POWER_EFFICIENT is
enabled.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>